Incorrect cleanup of payload
The payload, a proxy, was actually coming into the action creator via React.
The proxy was in fact a React SyntheticEvent, and when using the react-redux library,
this proxy is passed by default in the connect/mapDispatchToProps when not using
a callback, but providing an object literal of "connections".

Decided to keep it simple for now and just remove the ability to create anything
other than a default oscillator.